+++ library/ImboClient/Driver/DriverInterface.php
@@ -32,6 +32,10 @@
 
 namespace ImboClient\Driver;
 
+use ImboClient\Http\Response\ResponseInterface,
+    ImboClient\Exception\ServerException,
+    ImboClient\Exception\RuntimeException;
+
 /**
  * Client driver interface
  *
@@ -62,18 +66,19 @@ interface DriverInterface {
      *
      * @param string $url The URL to POST to
      * @param string $metadata The metadata to POST. The data must be JSON-encoded
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @param array $headers Additional headers to send in this request as an associative array
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
-    function post($url, $metadata);
+    function post($url, $metadata, array $headers = array());
 
     /**
      * PUT a file to an URL
      *
      * @param string $url The URL to PUT to
      * @param string $filePath Path to the local file
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
     function put($url, $filePath);
 
@@ -82,17 +87,18 @@ interface DriverInterface {
      *
      * @param string $url The URL to PUT to
      * @param string $data The data to PUT
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @param array $headers Additional headers to send in this request as an associative array
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
-    function putData($url, $data);
+    function putData($url, $data, array $headers = array());
 
     /**
      * Perform a GET to $url
      *
      * @param string $url The URL to GET
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
     function get($url);
 
@@ -100,8 +106,8 @@ interface DriverInterface {
      * Perform a HEAD to $url
      *
      * @param string $url The URL to HEAD
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
     function head($url);
 
@@ -109,25 +115,25 @@ interface DriverInterface {
      * Perform a DELETE request to $url
      *
      * @param string $url The URL to DELETE
-     * @return ImboClient\Http\Response\ResponseInterface
-     * @throws RuntimeException
+     * @return ResponseInterface
+     * @throws RuntimeException|ServerException
      */
     function delete($url);
 
     /**
-     * Add a request header
+     * Set a request header
      *
      * @param string $key The header key
      * @param string $value The value to send
-     * @return ImboClient\Driver\DriverInterface
+     * @return DriverInterface
      */
-    function addRequestHeader($key, $value);
+    function setRequestHeader($key, $value);
 
     /**
-     * Add one or more request headers
+     * Set one or more request headers
      *
      * @param array $headers Associative array
-     * @return ImboClient\Driver\DriverInterface
+     * @return DriverInterface
      */
-    function addRequestHeaders(array $headers);
+    function setRequestHeaders(array $headers);
 }